step 16 update DTO, create documentation

+++ src/main/java/com/example/ekaterinabeidel/service/CarService.java
@@ -1,10 +1,17 @@
 package com.example.ekaterinabeidel.service;
 
-import com.example.ekaterinabeidel.dto.CarDTO;
+import com.example.ekaterinabeidel.dto.CarCreateDTO;
+import com.example.ekaterinabeidel.dto.CarResponseDTO;
+import com.example.ekaterinabeidel.dto.CarUpdateDTO;
 import com.example.ekaterinabeidel.entity.Car;
+import com.example.ekaterinabeidel.entity.Engineer;
+import com.example.ekaterinabeidel.exception.CarNotFoundException;
+import com.example.ekaterinabeidel.exception.EngineerNotFoundException;
 import com.example.ekaterinabeidel.repository.CarRepository;
+import com.example.ekaterinabeidel.repository.EngineerRepository;
 import com.example.ekaterinabeidel.util.CarMapper;
 import com.example.ekaterinabeidel.util.EngineerMapper;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -16,41 +23,57 @@ public class CarService {
     @Autowired
     private CarRepository carRepository;
 
-    public List<CarDTO> getAllCars() {
-        return CarMapper.mapToListDTO(carRepository.findAll());
+    @Autowired
+    private EngineerRepository engineerRepository;
+
+    public List<CarResponseDTO> getAllCars() {
+        return CarMapper.mapToListCarResponseDTO(carRepository.findAll());
     }
 
-    public List<CarDTO> findByEngineerId(Long engineerId) {
-        return CarMapper.mapToListDTO(carRepository.findByEngineerId(engineerId));
+    public List<CarResponseDTO> findByEngineerId(Long engineerId) {
+        return CarMapper.mapToListCarResponseDTO(carRepository.findByEngineerId(engineerId));
     }
 
-    public List<CarDTO> findByBrandAndModel(String brand, String model) {
-        return CarMapper.mapToListDTO(carRepository.findByBrandAndModel(brand, model));
+    public List<CarResponseDTO> findByBrandAndModel(String brand, String model) {
+        return CarMapper.mapToListCarResponseDTO(carRepository.findByBrandAndModel(brand, model));
     }
 
-    public List<CarDTO> findByYearGreaterThan(int year) {
-        return CarMapper.mapToListDTO(carRepository.findByYearGreaterThan(year));
+    public List<CarResponseDTO> findByYearGreaterThan(int year) {
+        return CarMapper.mapToListCarResponseDTO(carRepository.findByYearGreaterThan(year));
     }
 
-    public CarDTO save(CarDTO carDTO) {
-        return CarMapper.mapToDTO(carRepository.save(CarMapper.mapToEntity(carDTO)));
-    }
-
-    public CarDTO updateCar(Long id, CarDTO carDTO) {
-        Optional<Car> optionalCar = carRepository.findById(id);
-        if (optionalCar.isPresent()) {
-            Car car = optionalCar.get();
-            car.setModel(carDTO.getModel());
-            car.setBrand(carDTO.getBrand());
-            car.setYear(carDTO.getYear());
-            car.setCityOfManufacture(carDTO.getCityOfManufacture());
-            car.setEngineer(EngineerMapper.mapToEntity(carDTO.getEngineerDTO()));
-            return CarMapper.mapToDTO(carRepository.save(car));
+    public CarResponseDTO save(CarCreateDTO carCreateDTO) {
+        Car car = CarMapper.mapCarCreateDTOToEntity(carCreateDTO);
+        if (carCreateDTO.getEngineerId() != null) {
+            Engineer engineer = engineerRepository.findById(carCreateDTO.getEngineerId())
+                    .orElseThrow(() -> new EntityNotFoundException("Engineer with ID " + carCreateDTO.getEngineerId() + " not found"));
+            car.setEngineer(engineer);
         }
-        return null;
+        Car savedCar = carRepository.save(car);
+        return CarMapper.mapToCarResponseDTO(savedCar);
+    }
+
+    public CarResponseDTO updateCar(Long id, CarUpdateDTO carUpdateDTO) {
+        Car car = carRepository.findById(id).orElseThrow(
+                () -> new CarNotFoundException("Car with ID " + id + " not found")
+        );
+        Engineer engineer = engineerRepository.findById(carUpdateDTO.getEngineerId()).orElseThrow(
+                () -> new EngineerNotFoundException("Engineer with ID " + carUpdateDTO.getEngineerId() + " not found")
+        );
+        car.setModel(carUpdateDTO.getModel());
+        car.setBrand(carUpdateDTO.getBrand());
+        car.setYear(carUpdateDTO.getYear());
+        car.setCityOfManufacture(carUpdateDTO.getCityOfManufacture());
+        car.setEngineer(engineer);
+        Car updatedCar = carRepository.save(car);
+        return CarMapper.mapToCarResponseDTO(updatedCar);
     }
 
     public void delete(Long id) {
-        carRepository.deleteById(id);
+        if (!carRepository.existsById(id)) {
+            throw new CarNotFoundException("Car with ID " + id + " not found");
+        } else {
+            carRepository.deleteById(id);
+        }
     }
 }
